+++ Source/UI/Theme.vb
@@ -476,7 +476,7 @@ Public Class Theme
 
         Public Function Clone() As ControlsThemeColors
             Using stream As Stream = New MemoryStream()
-                Dim formatter As IFormatter = New BinaryFormatter()
+                Dim formatter As IFormatter = SafeSerialization.CreateFormatter()
                 formatter.Serialize(stream, Me)
                 stream.Seek(0, SeekOrigin.Begin)
                 Return DirectCast(formatter.Deserialize(stream), ControlsThemeColors)
@@ -1028,6 +1028,5 @@ Public Class Theme
     End Class
 
 End Class
-